refactor!(value): add property and timestamp to the Value enum

+++ src/connection/incoming.rs
@@ -16,7 +16,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use tracing::{debug, warn};
+use tracing::{debug, instrument, warn};
 
 use crate::client::RecvError;
 use crate::error::{AggregationError, InterfaceTypeError};
@@ -32,6 +32,7 @@ impl<C> DeviceConnection<C>
 where
     C: Connection,
 {
+    #[instrument(skip(self, payload))]
     async fn handle_event(
         &self,
         interface: &str,
@@ -52,7 +53,7 @@ where
             }));
         };
 
-        let (data, timestamp) = match (interface.aggregation(), interface.interface_type()) {
+        let data = match (interface.aggregation(), interface.interface_type()) {
             (Aggregation::Individual, InterfaceTypeDef::Properties) => {
                 self.handle_property(interface, &path, payload).await?
             }
@@ -67,7 +68,7 @@ where
             }
         };
 
-        debug!("received {{v: {data:?}, t: {timestamp:?}}}");
+        debug!(?data, "received event");
 
         Ok(data)
     }
@@ -78,7 +79,7 @@ where
         interface: &Interface,
         path: &MappingPath<'_>,
         payload: C::Payload,
-    ) -> Result<(Value, Option<chrono::DateTime<chrono::Utc>>), TransportError>
+    ) -> Result<Value, TransportError>
     where
         C: Receive + Sync,
     {
@@ -114,9 +115,16 @@ where
                     interface.version_major()
                 );
 
-                Ok((Value::Individual(value), None))
+                Ok(Value::Property(Some(value)))
             }
             None => {
+                if !prop.allow_unset() {
+                    return Err(TransportError::Recv(RecvError::Unset {
+                        interface_name: prop.interface().interface_name().to_string(),
+                        path: prop.path().to_string(),
+                    }));
+                }
+
                 // Unset can only be received for a property
                 self.store
                     .delete_prop(&PropertyMapping::new_unchecked(
@@ -132,7 +140,7 @@ where
                     interface.version_major()
                 );
 
-                Ok((Value::Unset, None))
+                Ok(Value::Property(None))
             }
         }
     }
@@ -143,7 +151,7 @@ where
         interface: &Interface,
         path: &MappingPath<'_>,
         payload: C::Payload,
-    ) -> Result<(Value, Option<chrono::DateTime<chrono::Utc>>), TransportError>
+    ) -> Result<Value, TransportError>
     where
         C: Receive + Sync,
     {
@@ -154,11 +162,16 @@ where
             }));
         };
 
-        let (value, timestamp) = self.connection.deserialize_individual(&mapping, payload)?;
+        let (data, timestamp) = self.connection.deserialize_individual(&mapping, payload)?;
 
-        // TODO: maybe move validation from each connection to here
+        let timestamp = Self::validate_timestamp(
+            interface.interface_name(),
+            path.as_str(),
+            mapping.explicit_timestamp(),
+            timestamp,
+        )?;
 
-        Ok((Value::Individual(value), timestamp))
+        Ok(Value::Individual { data, timestamp })
     }
 
     /// Handles the payload of an interface with [`InterfaceAggregation::Object`]
@@ -167,7 +180,7 @@ where
         interface: &Interface,
         path: &MappingPath<'_>,
         payload: C::Payload,
-    ) -> Result<(Value, Option<chrono::DateTime<chrono::Utc>>), TransportError>
+    ) -> Result<Value, TransportError>
     where
         C: Receive + Sync,
     {
@@ -183,7 +196,14 @@ where
 
         let (data, timestamp) = self.connection.deserialize_object(&object, path, payload)?;
 
-        Ok((Value::Object(data), timestamp))
+        let timestamp = Self::validate_timestamp(
+            interface.interface_name(),
+            path.as_str(),
+            object.explicit_timestamp(),
+            timestamp,
+        )?;
+
+        Ok(Value::Object { data, timestamp })
     }
 
     pub(super) async fn handle_connection_event(
